Move TimeBarChart to components/index/cards/charts
