+++ components/index/cards/charts/TimeBarChart.vue
